# knowledge/docs/api/api_spec.yaml
openapi: 3.0.3
info:
  title: 预测性智能交易系统 API
  description: |
    智能交易系统的 RESTful API 规范
    
    ## 认证方式
    所有 API 请求需要在 Header 中包含 JWT Token:
    ```
    Authorization: Bearer <token>
    ```
    
    ## 速率限制
    - 普通端点：100 请求/分钟
    - 交易端点：20 请求/分钟
    - 数据端点：200 请求/分钟
    
    ## 错误码
    - 400: 请求参数错误
    - 401: 未授权
    - 403: 禁止访问
    - 404: 资源不存在
    - 429: 请求超限
    - 500: 服务器内部错误
  version: 1.0.0
  contact:
    name: API Support
    email: support@trading-system.com
servers:
  - url: https://api.trading-system.com/v1
    description: 生产环境
  - url: https://staging-api.trading-system.com/v1
    description: 测试环境
  - url: http://localhost:8000/v1
    description: 本地开发
...
paths:
...
  /orders:
    post:
      tags:
        - 交易
      summary: 创建订单
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: 订单创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    
    get:
      tags:
        - 交易
      summary: 获取订单列表
      security:
        - BearerAuth: []
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: symbol
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, filled, cancelled, rejected]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/{order_id}:
    get:
      tags:
        - 交易
      summary: 获取订单详情
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    
    delete:
      tags:
        - 交易
      summary: 取消订单
      security:
        - BearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 取消成功
  /positions:
    get:
      tags:
        - 交易
      summary: 获取当前仓位列表
      security:
        - BearerAuth: []
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: symbol
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
  /positions/{position_id}:
    get:
      tags:
        - 交易
      summary: 获取仓位详情
      security:
        - BearerAuth: []
      parameters:
        - name: position_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Position'

  # ==================== 风控端点 ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        account_id:
          type: string
          format: uuid
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [market, limit, stop_limit, trigger]
        size:
          type: number
        price:
          type: number
        filled_size:
          type: number
        avg_fill_price:
          type: number
        status:
          type: string
          enum: [pending, partially_filled, filled, cancelled, rejected]
        time_in_force:
          type: string
          enum: [GTC, IOC, FOK]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required:
        - account_id
        - symbol
        - side
        - type
        - size
      properties:
        account_id:
          type: string
          format: uuid
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [market, limit, stop_limit, trigger]
        size:
          type: number
        price:
          type: number
        stop_price:
          type: number
        time_in_force:
          type: string
          enum: [GTC, IOC, FOK]
        leverage:
          type: number
        reduce_only:
          type: boolean
          default: false
...
    Position:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trade_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        symbol:
          type: string
        side:
          type: string
          enum: [LONG, SHORT]
        size:
          type: number
        entry_price:
          type: number
        current_price:
          type: number
        mark_price:
          type: number
        unrealized_pnl:
          type: number
        unrealized_pnl_percent:
          type: number
        leverage:
          type: number
        margin:
          type: number
        liquidation_price:
          type: number
        stop_loss:
          type: number
        take_profit:
          type: array
          items:
            type: number
        opened_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...